Stop frames from allowing itself to navigate top without gesture
A previous CL stopped a sandboxed iframe from navigating top if its
ancestor was not allowed to navigate top[1]. However, there is still a
corner case that allows a sandboxed frame to give itself permission to
navigate top if the top-level page doesn't have any sandbox flags set.

This is problematic because it allows a cross-origin iframe to
circumvent our framebust intervention and redirect the top page to a
potentially malicious page.

The fix:
If the following conditions hold:
- The embedder is unsandboxed
- The frame being navigated is cross-origin to its embedder
- The frame is sandboxed
- The frame's sandbox flags are set to `allow-top-navigation`
The following intervention will happen:
- Switch the frame's `allow-top-navigation` flag to
  `allow-top-navigation-by-user-activation`
- Output a warning message to the console explaining the intervention.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2688360

Bug: 1251790
Change-Id: I7406d631d6b9c4bdbfc71c433db50b2fca2f0c21