[CompositeClipPathAnimations] Fix crash caused by position: fixed.
With a translation of 0,0, a DCHECK in
DrawingDisplayItem::CalculateRectKnownToBeOpaqueForRecord would fail, as
the 'image size' for a deferred image is set by a parameter in the paint
worklet input, which in this case was the reference box. This would
fail the DCHECK when the clip path was larger than the reference box.

An extra parameter has been added to the clip path paint worklet input
so that the reference box is independent from the size of the clip area.

The area painted when producing the deferred image has also been
translated so that its origin is 0,0. This is also necessary to avoid the
DCHECK.

Bug: 1353038
Change-Id: I1d62a00a92f38e06e2ae090caefced5fdf289fd9