CacheStorage now keeps the fragment in the Request URL but doesn't save it in the Cache key.
The Cache storage now keeps in the Request Metadata, the complete URL (including the fragment). However, when doing the match, the fragment part is removed. This means that these 3 example URLs will all match:

https://example.com/path
https://example.com/path#ref
https://example.com/path#ref2

The latest url put in the cache will be the one saved and returned in the Request.

This CL modified the following:

- `third_party/blink/renderer/core/fetch/request.h`

The `removeFragment` parameter was added to the `CreateFetchAPIRequest(
      bool removeFragment = true)` and by default, for retro-compatibility reasons, the fragment is removed.

Once we are certain that always keeping the fragment doesn't cause any side-effects, we can change this behaviour and just remove the code that was removing the fragment.

- Methods to support manipulating fragments (ref) and their unit tests were also added.

- The cache now saves the url in the request metadata so modifications to the proto file were made.

Bug:1205111
Change-Id: Ib23fd5a9364bc64159a2859dc9fd9d3512697202