Consolidate Timing-Allow-Origin checking code
Actual TAO check is done in a function TimingAllowOriginCheck()

blink/ResourceTiming doesn't check for TAO anymore.
network::cors::CorsUrlLoaded performs the checks in transit, as
specified in https://fetch.spec.whatwg.org/#tao-check, which now also
applies to nested contexts (iframes).

The exception is the resource timing for a cross-origin iframe served
from a service worker - the TAO check for that resource is done in
ServiceWorkerMainResourceLoaded as loading of a main resource in a
service worker doesn't have the same context to decide about TAO as
CorsUrLLoader.
See resource-timing/workerStart-tao-protected.https.html

resource-timing-cross-origin.https.html now works according to spec,
amended it.

Added unit tests for CorsUrlLoader and TimingAllowOriginParser.

Fixed: 1201767
Bug: 1201767
Change-Id: Ic85884cb8400dbef9e8dba1409bfec3f0dd9538c