Navigation-rewrite: Add unparseable URL redirect traversal test
This CL adds a test for row 4 in our navigation and session history
rewrite test spreadsheet [1]. This test ensures that upon redirects,
the right URL is saved in the session history entry.

Specifically, if /A navigates to /B, which redirects to an unparseable
URL (which results in an error Document), we want to ensure that the
new session history entry's URL is /B, not /A or the unparseable URL
redirect.

[1]:
https://docs.google.com/spreadsheets/d/1uqSS7Lwm_y_bVYFJBFVrZG6YKwfyKPGrH0ZKJIKwwMY/edit#gid=0.

R=domenic@chromium.org

Bug: N/A
Change-Id: I07d2a5d0e193cc62a7115bae2466a96fbc817fc8